Version 1.6.0
### Features
- Update `IHazExample` and `ICompileExample` (Fix: #59)
### Updated
- Change `IHazExample` Name to wildcard pattern `*.Example`
- Move `ISign` methods to `IHazSign`
- Change `IHazRelease` Folder to ReleaseFolder
