+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] / 2023-10-02
+### Features
+- Update `IHazExample` and `ICompileExample` (Fix: #59)
+### Updated
+- Change `IHazExample` Name to wildcard pattern `*.Example`
+- Move `ISign` methods to `IHazSign`
+- Change `IHazRelease` Folder to ReleaseFolder
+
 ## [1.5.2] / 2023-10-02
 ### Features
 - TestResult with `TargetFramework` in name. (Fix: #60)
@@ -277,6 +285,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.6.0]: ../../compare/1.5.2...1.6.0
 [1.5.2]: ../../compare/1.5.1...1.5.2
 [1.5.1]: ../../compare/1.5.0...1.5.1
 [1.5.0]: ../../compare/1.4.4...1.5.0
